Move NativeLibraryKind to rustc_session
